fixed some bugs about the composite pattern and updated class test work ended for today
have a nice night ;)

+++ MP/src/item/TVSet.java
@@ -2,14 +2,14 @@ package item;
 
 public class TVSet extends ConcreteItem {
 // classe leaf del pattern composite	
-	private Float inches;
+	private Integer inches;
 
-	public TVSet(String nm, String br, Float pr, String desc, Float in) {
+	public TVSet(String nm, String br, Float pr, String desc, Integer in) {
 		super(br, nm, desc, pr);
 		inches = in;
 	}
 
-	public Float getInches() {return inches;}
+	public Integer getInches() {return inches;}
 	
 
 }
